Make the user detail section a component
