+++ frontend/components/Layout/NavBars/AdminNavBar.tsx
@@ -19,31 +19,13 @@ import { HttpException } from "util/HttpExceptions";
 import { useUser } from "util/UserContext";
 import { createNewPage } from "util/api/pageApi";
 import NavBar, { Routes } from "./NavBar";
+import UserDetailsSection from "./UserDetailSection";
 
 type AdminNavBar = {
   userDetails: UserDetails;
   routes: Routes[];
   courseData?: UserCourseInformation;
   showAddPage?: boolean;
-};
-
-/**
- * User avatar
- */
-const UserDetailSection = ({ first_name, last_name, role, avatar }: UserDetails): JSX.Element => {
-  return (
-    <div className="mt-5 flex flex-row justify-center">
-      <div className="w-[50px] h-[50px] bg-orange-500 rounded-full flex justify-center items-center">
-        <span className="text-2xl font-bold">
-          {(first_name?.charAt(0) ?? "") + (last_name?.charAt(0) ?? "")}
-        </span>
-      </div>
-      <div className="flex flex-col pl-2 justify-center items-center">
-        <span className="font-bold text-start w-full">{`${first_name} ${last_name}`}</span>
-        <span className="text-start w-full">{getRoleText(role)}</span>
-      </div>
-    </div>
-  );
 };
 
 type CourseDetailsProps = {
@@ -163,7 +145,7 @@ export default function AdminNavBar({
         <div className="w-full flex flex-col justify-between h-[calc(100%_-_4rem)]">
           <div>
             {/* Top */}
-            {courseData === undefined && <UserDetailSection {...userDetails} />}
+            {courseData === undefined && <UserDetailsSection {...userDetails} />}
             {courseData !== undefined && <CourseDetails code={courseData?.code ?? ""} />}
             <NavBar
               routes={dynamicRoutes}
